static/padding.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding">
<meta name="assert" content="scroll-padding supports the full grammar '[ <length-percentage> | auto ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-padding-top", "auto");
test_valid_value("scroll-padding-top", "10px");
test_valid_value("scroll-padding-top", "20%");
test_valid_value("scroll-padding-top", "calc(2em + 3ex)");
test_valid_value("scroll-padding-top", "calc(50% + 60px)");


test_valid_value("scroll-padding-right", "auto");
test_valid_value("scroll-padding-right", "10px");
test_valid_value("scroll-padding-right", "20%");
test_valid_value("scroll-padding-right", "calc(2em + 3ex)");
test_valid_value("scroll-padding-right", "calc(50% + 60px)");


test_valid_value("scroll-padding-bottom", "auto");
test_valid_value("scroll-padding-bottom", "10px");
test_valid_value("scroll-padding-bottom", "20%");
test_valid_value("scroll-padding-bottom", "calc(2em + 3ex)");
test_valid_value("scroll-padding-bottom", "calc(50% + 60px)");


test_valid_value("scroll-padding-left", "auto");
test_valid_value("scroll-padding-left", "10px");
test_valid_value("scroll-padding-left", "20%");
test_valid_value("scroll-padding-left", "calc(2em + 3ex)");
test_valid_value("scroll-padding-left", "calc(50% + 60px)");


test_valid_value("scroll-padding", "auto");
test_valid_value("scroll-padding", "10px");
test_valid_value("scroll-padding", "20%");
test_valid_value("scroll-padding", "calc(2em + 3ex)");

test_valid_value("scroll-padding", "1px 2px");
test_valid_value("scroll-padding", "1px 2px 3%");
test_valid_value("scroll-padding", "1px 2px 3% 4px");
test_valid_value("scroll-padding", "1px auto");
test_valid_value("scroll-padding", "auto auto auto auto");






test_valid_value("scroll-padding-block-start", "auto");
test_valid_value("scroll-padding-block-start", "10px");
test_valid_value("scroll-padding-block-start", "20%");
test_valid_value("scroll-padding-block-start", "calc(2em + 3ex)");
test_valid_value("scroll-padding-block-start", "calc(50% + 60px)");


test_valid_value("scroll-padding-block-end", "auto");
test_valid_value("scroll-padding-block-end", "10px");
test_valid_value("scroll-padding-block-end", "20%");
test_valid_value("scroll-padding-block-end", "calc(2em + 3ex)");
test_valid_value("scroll-padding-block-end", "calc(50% + 60px)");


test_valid_value("scroll-padding-block", "auto");
test_valid_value("scroll-padding-block", "10px");
test_valid_value("scroll-padding-block", "20%");
test_valid_value("scroll-padding-block", "calc(2em + 3ex)");
test_valid_value("scroll-padding-block", "calc(50% + 60px)");

test_valid_value("scroll-padding-block", "1px 2px");
test_valid_value("scroll-padding-block", "1px auto");
test_valid_value("scroll-padding-block", "auto auto");





test_valid_value("scroll-padding-inline-start", "auto");
test_valid_value("scroll-padding-inline-start", "10px");
test_valid_value("scroll-padding-inline-start", "20%");
test_valid_value("scroll-padding-inline-start", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline-start", "calc(50% + 60px)");


test_valid_value("scroll-padding-inline-end", "auto");
test_valid_value("scroll-padding-inline-end", "10px");
test_valid_value("scroll-padding-inline-end", "20%");
test_valid_value("scroll-padding-inline-end", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline-end", "calc(50% + 60px)");


test_valid_value("scroll-padding-inline", "auto");
test_valid_value("scroll-padding-inline", "10px");
test_valid_value("scroll-padding-inline", "20%");
test_valid_value("scroll-padding-inline", "calc(2em + 3ex)");
test_valid_value("scroll-padding-inline", "calc(50% + 60px)");

test_valid_value("scroll-padding-inline", "1px 2px");
test_valid_value("scroll-padding-inline", "1px auto");
test_valid_value("scroll-padding-inline", "auto auto");
</script>
</body>
</html>
